Attempt to break out of cycle import dependency.
